Restructure type interfaces: created new ArrayType and ClassType interfaces and gutted the Type interface.  Clarified how ParameterizedType works; also renamed getRawClass() to getBaseType().  Moved typeParameters() from ClassDoc to ClassType.  Added a type() method to ClassDoc to get its ClassType.  Tweaked API types to accomodate this refactoring, where necessary.
Updated the implementation in net.cscott.gjdoc.parser to reflect the interface
changes.  Implemented PArrayType, PTypeVariable, PClassTypeVariable, and
PMethodTypeVariable.  Resurrected PParameterizedType, killed
PLazyParameterizedType and PEagerParameterizedType.

PLazyClassType has a slight issue in that TypeContext looks up a Type --
what if the name stored by PLazyClassType actually ends up referring to
a type *variable*?  Should probably refactor a little more to fix this.
At the moment, there is an unsafe cast to ClassType in PLazyClassType.lookup()
to finesse the point.