Refactor PTag *again*.  This time I like how it hangs together. Tag constructors throw TagParseException to indicate a problem parsing, in which case an opaque tag type is returned, instead of the parsed subclass type.  One potential disadvantage: you must be careful to test the type of the Tag object, and not rely on the tag name to tell you if a tag is (for example) a PParamTag.
Other niggling detail still unresolved: tag parsing happens lazily,
when the 'tags()' method is called, and is not cached.  So the
error message occur at unusual times, and may repeat themselves.
A quick fix is to cache the result of PDoc.tags().  A longer-term
solution would be to eagerly parse the tags, but it's hard to do this
in PDoc considering the constraints on superclass initialization ordering.