Two bug fixes to HTMLUtil, both having to do with HTMLWriter.  First, we can't share a CharsetEncoder instance between the HTMLWriter and an OutputStreamWriter, or eventually the HTMLWriter will screw up the state of an on-going encoding operation in the OuputStreamWriter.
Secondly, add a little bit of safely in case we're called with
len==0.  The 'canEncode' method *should* always return true in this
case, in which case we're okay -- but if it ever returned false we'd
potentially index an array off its end and crash.  So be safe, and
explicitly handle the len==0 case at the top of the method.