Revamp type-parameter related macros, splitting into three types: TYPEPARAMS_SUMMARY, which just emits the type variable names without hyperlinks or bounds; TYPEPARAMS_LINK which emits the variables with hyperlinks and bounds; and TYPEPARAMS_DECL which emits the variables with bounds as anchors.
Most of the implementation moved to HTMLUtil.toLink().    A
TemplateContext.specificTypeVariables() method was added as well, to
return the most specific type variables in the context.

The class-page.html template was modified to use these macros appropriately.

This commit fixes the bug that we didn't actually emit type parameter
bounds in the HTML output.  Still left to fix:
  class Enum<E extends Enum<E>>
in this declaration the final 'E' is not correctly parsed as referring to
the type parameter at the head of the bound.  It is parsed as referring to
some unknown class E.  Similarly for:
   static <E extends Enum<E>> E valueOf(List<E> family, String name)
in this class, where the E in the bound is parsed as referring to the
*class variable* E, not the *method type variable*.  Also, the TypeContext
for static members shouldn't include the enclosing class.

Nevertheless, I'm calling this version 0.1 final.  I'll fix these bugs
in version 0.2.