"GJDoc" is already the name of the pizza/GJ documentation tool    http://www.cis.unisa.edu.au/~pizza/gj/ as well as a GNU version of javadoc    http://www.gnu.org/software/cp-tools/gjdoc.en.html
I've no wish to further confuse the issue.  So my tool is henceforth
renamed 'SinjDoc', in the grand tradition of recursive acronyms for
software projects, for 'SinjDoc Is Not JavaDoc' -- it's better!

I've replaced GJDoc wherever it appears in the documentation, etc,
but the package name is still 'net.cscott.gjdoc'.  I'll fix that
in the next commit.